Improved reliability of TCP tx
Found that a crash was being caused by running out of memory. We were
only raising idle priority to free deleted task resources when we failed
to create a task, but the same thing that caused task creation to fail
was also causing a crash when we tried to alloc a new message.

Also removed the allocs from Dequeue, and it now takes a pointer