interactive assembly checkpoint
working:
press button on module a to set as input. hold button to change to
output. module b gets messages telling it to configure itself as the
opposite of module a. module b stores module a's address and sends back
an accept message. module a stores module b's address. when module a
times out, it sends a message to module b to cancel.

not working:
actually sending data over the channel
keeping the channel alive