Fixed crash on ESP startup. sped up start by eliminating IMU init (for now) and moving the TCP init up in the order.
Moving TCP init seems to have helped the speed with which the ESP finds
a connection too, as we are no longer sending it a bunch of messages
while it's trying to connect. With this firmware, you can reliably start
linking channels with button presses about 10 seconds after powering on
the board.